restructure to speparate edgs.
